requirements.txt
ansible==2.9.5
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
paramiko==2.7.1
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.3.5
PyYAML==5.3
six==1.14.0
testinfra==4.0.0
wcwidth==0.1.8
